feat: actionbox restructure
