--- apps/marginfi-v2-ui/src/components/common/ActionBox/ActionBoxLend/index.ts
@@ -1,1 +1,0 @@
-export * from './ActionBoxAvailableCollateral'
