+++ apps/marginfi-v2-ui/src/components/common/ActionBox/ActionBoxPreview/index.ts
@@ -1,0 +1,1 @@
+export * from "./ActionBoxPreview";
