+++ apps/marginfi-v2-ui/src/components/common/ActionBox/index.tsx
@@ -2,6 +2,6 @@ export * from "./ActionBox.utils";
 export * from "./ActionBox";
 export * from "./ActionBoxActions";
 export * from "./ActionBoxDialog";
-export * from "./ActionBoxPreview";
 export * from "./ActionBoxPriorityFees";
 export * from "./ActionBoxTokens";
+export * from "./ActionBoxLendWrapper";
