+++ apps/marginfi-v2-ui/src/utils/featureGates.ts
@@ -4,7 +4,7 @@ enum Features {
 
 function isActive(feature: Features) {
   const featureGatesRaw = process.env.NEXT_PUBLIC_FEATURE_GATES as string | undefined;
-  if (!featureGatesRaw) return false;
+  if (!featureGatesRaw) return true;
 
   const featureGates = JSON.parse(featureGatesRaw) as Record<string, boolean>;
